+++ src/cartesian/Area.js
@@ -63,14 +63,16 @@ class Area extends Component {
       y: PropTypes.number,
       value: PropTypes.oneOfType([PropTypes.number, PropTypes.array]),
     })),
+
     onAnimationStart: PropTypes.func,
     onAnimationEnd: PropTypes.func,
-
     animationId: PropTypes.number,
     isAnimationActive: PropTypes.bool,
     animationBegin: PropTypes.number,
     animationDuration: PropTypes.number,
     animationEasing: PropTypes.oneOf(['ease', 'ease-in', 'ease-out', 'ease-in-out', 'linear']),
+
+    id: PropTypes.string,
   };
 
   static defaultProps = {
@@ -358,8 +360,9 @@ class Area extends Component {
 
   renderAreaWithAnimation(needClip) {
     const { points, baseLine, isAnimationActive, animationBegin,
-      animationDuration, animationEasing, animationId } = this.props;
+      animationDuration, animationEasing, animationId, id } = this.props;
     const { prevPoints, prevBaseLine } = this.state;
+    const clipPathId = _.isNil(id) ? this.id : id;
 
     return (
       <Animate
@@ -393,6 +396,9 @@ class Area extends Component {
               if (isNumber(baseLine)) {
                 const interpolator = interpolateNumber(prevBaseLine, baseLine);
                 stepBaseLine = interpolator(t);
+              } else if (_.isNil(baseLine) || _.isNaN(baseLine)) {
+                const interpolator = interpolateNumber(prevBaseLine, 0);
+                stepBaseLine = interpolator(t);
               } else {
                 stepBaseLine = baseLine.map((entry, index) => {
                   if (prevBaseLine[index]) {
@@ -413,11 +419,11 @@ class Area extends Component {
             return (
               <Layer>
                 <defs>
-                  <clipPath id={`animationClipPath-${this.id}`}>
+                  <clipPath id={`animationClipPath-${clipPathId}`}>
                     {this.renderClipRect(t)}
                   </clipPath>
                 </defs>
-                <Layer clipPath={`url(#animationClipPath-${this.id})`}>
+                <Layer clipPath={`url(#animationClipPath-${clipPathId})`}>
                   {this.renderAreaStatically(points, baseLine, needClip)}
                 </Layer>
               </Layer>
@@ -443,7 +449,7 @@ class Area extends Component {
 
   render() {
     const { hide, dot, points, className, top, left, xAxis, yAxis,
-      width, height, isAnimationActive } = this.props;
+      width, height, isAnimationActive, id } = this.props;
 
     if (hide || !points || !points.length) { return null; }
 
@@ -451,12 +457,13 @@ class Area extends Component {
     const hasSinglePoint = points.length === 1;
     const layerClass = classNames('recharts-area', className);
     const needClip = (xAxis && xAxis.allowDataOverflow) || (yAxis && yAxis.allowDataOverflow);
+    const clipPathId = _.isNil(id) ? this.id : id;
 
     return (
       <Layer className={layerClass}>
         {needClip ? (
           <defs>
-            <clipPath id={`clipPath-${this.id}`}>
+            <clipPath id={`clipPath-${clipPathId}`}>
               <rect x={left} y={top} width={width} height={height} />
             </clipPath>
           </defs>