+++ src/chart/RadarChart.js
@@ -18,8 +18,8 @@ export default generateCategoricalChart({
   formatAxisMap,
   defaultProps: {
     layout: 'centric',
-    startAngle: 360,
-    endAngle: 0,
+    startAngle: 90,
+    endAngle: -270,
     cx: '50%',
     cy: '50%',
     innerRadius: 0,