+++ src/chart/generateCategoricalChart.js
@@ -19,13 +19,13 @@ import { findAllByType, findChildByType, getDisplayName, parseChildIndex,
 import CartesianAxis from '../cartesian/CartesianAxis';
 import Brush from '../cartesian/Brush';
 import { getOffset, calculateChartCoordinate } from '../util/DOMUtils';
-import { getAnyElementOfObject, hasDuplicate, uniqueId, isNumber } from '../util/DataUtils';
+import { getAnyElementOfObject, hasDuplicate, uniqueId, isNumber, findEntryInArray } from '../util/DataUtils';
 import { calculateActiveTickIndex, getMainColorOfGraphicItem, getBarSizeList,
   getBarPosition, appendOffsetOfLegend, getLegendProps, combineEventHandlers,
   getTicksOfAxis, getCoordinatesOfGrid, getStackedDataOfItem, parseErrorBarsOfAxis,
   getBandSizeOfAxis, getStackGroupsByAxisId, getValueByDataKey, isCategorialAxis,
   getDomainOfItemsWithSameAxis, getDomainOfStackGroups, getDomainOfDataByKey,
-  detectReferenceElementsDomain, parseSpecifiedDomain } from '../util/ChartUtils';
+  detectReferenceElementsDomain, parseSpecifiedDomain, parseDomainOfCategoryAxis } from '../util/ChartUtils';
 import { inRangeOfSector, polarToCartesian } from '../util/PolarUtils';
 import { shallowEqual } from '../util/PureRender';
 import { eventCenter, SYNC_EVENT } from '../util/Events';
@@ -76,6 +76,7 @@ const generateCategoricalChart = ({
       onMouseDown: PropTypes.func,
       onMouseUp: PropTypes.func,
       reverseStackOrder: PropTypes.bool,
+      id: PropTypes.string,
       ...propTypes,
     };
 
@@ -155,7 +156,7 @@ const generateCategoricalChart = ({
       this.state = { ...defaultState, updateId: 0,
         ...this.updateStateOfAxisMapsOffsetAndStackGroups({ props, ...defaultState, updateId }) };
 
-      this.uniqueChartId = uniqueId('recharts');
+      this.uniqueChartId = _.isNil(props.id) ? uniqueId('recharts') : props.id;
 
       if (props.throttleDelay) {
         this.triggeredAfterMouseMove = _.throttle(this.triggeredAfterMouseMove,
@@ -183,6 +184,7 @@ const generateCategoricalChart = ({
             { props: nextProps, ...defaultState, updateId: updateId + 1 }) }
         );
       } else if (!isChildrenEqual(nextProps.children, children)) {
+        // update configuration in chilren
         const hasGlobalData = !_.isNil(nextProps.data);
         const newUpdateId = hasGlobalData ? updateId : updateId + 1;
         const { dataStartIndex, dataEndIndex } = this.state;
@@ -190,6 +192,7 @@ const generateCategoricalChart = ({
         const defaultState = {
           ...this.constructor.createDefaultState(nextProps), dataEndIndex, dataStartIndex,
         };
+
         this.setState({
           ...defaultState,
           updateId: newUpdateId,
@@ -267,7 +270,8 @@ const generateCategoricalChart = ({
 
       // Eliminate duplicated axes
       const axisMap = axes.reduce((result, child) => {
-        const { type, dataKey, allowDataOverflow, scale, ticks } = child.props;
+        const { type, dataKey, allowDataOverflow, allowDuplicatedCategory,
+          scale, ticks } = child.props;
         const axisId = child.props[axisIdKey];
         const displayedData = this.constructor.getDisplayedData(props, {
           graphicalItems: graphicalItems.filter(item => item.props[axisIdKey] === axisId),
@@ -284,13 +288,29 @@ const generateCategoricalChart = ({
 
             if (type === 'category' && isCategorial) {
               const duplicate = hasDuplicate(domain);
-              duplicateDomain = duplicate ? domain : null;
 
-              // When category axis has duplicated text, serial numbers are used to generate scale
-              domain = duplicate ? _.range(0, len) : domain;
+              if (allowDuplicatedCategory && duplicate) {
+                duplicateDomain = domain;
+                // When category axis has duplicated text, serial numbers are used to generate scale
+                domain = _.range(0, len);
+              } else if (!allowDuplicatedCategory) {
+                // remove duplicated category
+                domain = parseDomainOfCategoryAxis(child.props.domain, domain, child)
+                  .reduce((finalDomain, entry) => (
+                    finalDomain.indexOf(entry) >= 0 ? finalDomain : [...finalDomain, entry]
+                  ), []);
+              }
             } else if (type === 'category') {
-              // eliminate undefined or null or empty string
-              domain = domain.filter(entry => (entry !== '' && !_.isNil(entry)));
+              if (!allowDuplicatedCategory) {
+                domain = parseDomainOfCategoryAxis(child.props.domain, domain, child)
+                  .reduce((finalDomain, entry) => (
+                    (finalDomain.indexOf(entry) >= 0 || entry === '' || _.isNil(entry)) ?
+                      finalDomain : [...finalDomain, entry]
+                  ), []);
+              } else {
+                // eliminate undefined or null or empty string
+                domain = domain.filter(entry => (entry !== '' && !_.isNil(entry)));
+              }
             } else if (type === 'number') {
               const errorBarsDomain = parseErrorBarsOfAxis(
                 displayedData,
@@ -435,9 +455,7 @@ const generateCategoricalChart = ({
 
     getActiveCoordinate(tooltipTicks, activeIndex, rangeObj) {
       const { layout } = this.props;
-      const entry = _.get(tooltipTicks.filter(tick => (
-        tick && (tick.index === activeIndex)
-      )), '[0]');
+      const entry = tooltipTicks.find(tick => tick && (tick.index === activeIndex));
 
       if (entry) {
         if (layout === 'horizontal') {
@@ -497,7 +515,7 @@ const generateCategoricalChart = ({
 
       if (activeIndex >= 0 && tooltipTicks) {
         const activeLabel = tooltipTicks[activeIndex] && tooltipTicks[activeIndex].value;
-        const activePayload = this.getTooltipContent(activeIndex);
+        const activePayload = this.getTooltipContent(activeIndex, activeLabel);
         const activeCoordinate = this.getActiveCoordinate(ticks, activeIndex, rangeObj);
 
         return {
@@ -512,10 +530,11 @@ const generateCategoricalChart = ({
     /**
      * Get the content to be displayed in the tooltip
      * @param  {Number} activeIndex    Active index of data
+     * @param  {String} activeLabel    Active label of data
      * @return {Array}                 The content of tooltip
      */
-    getTooltipContent(activeIndex) {
-      const { graphicalItems } = this.state;
+    getTooltipContent(activeIndex, activeLabel) {
+      const { graphicalItems, tooltipAxis } = this.state;
       const displayedData = this.constructor.getDisplayedData(this.props, this.state);
 
       if (activeIndex < 0 || !graphicalItems || !graphicalItems.length
@@ -523,19 +542,30 @@ const generateCategoricalChart = ({
         return null;
       }
 
+      // get data by activeIndex when the axis don't allow duplicated category
       return graphicalItems.reduce((result, child) => {
         const { hide } = child.props;
+
         if (hide) { return result; }
 
-        const { dataKey, name, unit, formatter } = child.props;
+        const { dataKey, name, unit, formatter, data } = child.props;
+        let payload;
 
+        if (tooltipAxis.dataKey && !tooltipAxis.allowDuplicatedCategory) {
+          // graphic child has data props
+          payload = findEntryInArray(data || displayedData, tooltipAxis.dataKey, activeLabel);
+        } else {
+          payload = displayedData[activeIndex];
+        }
+
+        if (!payload) { return result; }
         return [...result, {
           ...getPresentationAttributes(child),
           dataKey, unit, formatter,
           name: name || dataKey,
           color: getMainColorOfGraphicItem(child),
-          value: getValueByDataKey(displayedData[activeIndex], dataKey),
-          payload: displayedData[activeIndex],
+          value: getValueByDataKey(payload, dataKey),
+          payload,
         }];
       }, []);
     }
@@ -907,13 +937,13 @@ const generateCategoricalChart = ({
       if (startIndex !== this.state.dataStartIndex || endIndex !== this.state.dataEndIndex) {
         const { updateId } = this.state;
 
-        this.setState({
+        this.setState(() => ({
           dataStartIndex: startIndex,
           dataEndIndex: endIndex,
           ...this.updateStateOfAxisMapsOffsetAndStackGroups(
             { props: this.props, dataStartIndex: startIndex, dataEndIndex: endIndex, updateId }
           ),
-        });
+        }));
 
         this.triggerSyncEvent({
           dataStartIndex: startIndex,
@@ -959,21 +989,21 @@ const generateCategoricalChart = ({
      * @return {Object} no return
      */
     handleItemMouseEnter = (el) => {
-      this.setState({
+      this.setState(() => ({
         isTooltipActive: true,
         activeItem: el,
         activePayload: el.tooltipPayload,
         activeCoordinate: el.tooltipPosition || { x: el.cx, y: el.cy },
-      });
+      }));
     };
     /**
      * The handler of mouse leaving a scatter
      * @return {Object} no return
      */
     handleItemMouseLeave = () => {
-      this.setState({
+      this.setState(() => ({
         isTooltipActive: false,
-      });
+      }));
     };
     /**
      * The handler of mouse moving in chart
@@ -1249,8 +1279,10 @@ const generateCategoricalChart = ({
 
       if (!props) { return null; }
 
-      return createElement(Legend, {
-        ...props,
+      const { item, ...otherProps } = props;
+
+      return cloneElement(item, {
+        ...otherProps,
         chartWidth: width,
         chartHeight: height,
         margin,
@@ -1376,25 +1408,34 @@ const generateCategoricalChart = ({
       if (!item) { return null; }
 
       const graphicalItem = cloneElement(element, item.props);
-      const { isTooltipActive, activeTooltipIndex } = this.state;
+      const { isTooltipActive, tooltipAxis, activeTooltipIndex, activeLabel } = this.state;
       const { children } = this.props;
       const tooltipItem = findChildByType(children, Tooltip);
       const { points, isRange, baseLine } = item.props;
       const { activeDot, hide } = item.item.props;
       const hasActive = !hide && isTooltipActive && tooltipItem && activeDot &&
-        activeTooltipIndex >= 0 && points[activeTooltipIndex];
+        activeTooltipIndex >= 0;
 
       if (hasActive) {
-        const activePoint = points[activeTooltipIndex];
-        const basePoint = isRange && baseLine && baseLine[activeTooltipIndex];
+        let activePoint, basePoint;
 
-        return [
-          graphicalItem,
-          ...this.renderActivePoints({
-            item, activePoint, basePoint, childIndex: activeTooltipIndex,
-            isRange,
-          }),
-        ];
+        if (tooltipAxis.dataKey && !tooltipAxis.allowDuplicatedCategory) {
+          activePoint = findEntryInArray(points, `payload.${tooltipAxis.dataKey}`, activeLabel);
+          basePoint = isRange && baseLine && findEntryInArray(baseLine, `payload.${tooltipAxis.dataKey}`, activeLabel);
+        } else {
+          activePoint = points[activeTooltipIndex];
+          basePoint = isRange && baseLine && baseLine[activeTooltipIndex];
+        }
+
+        if (!_.isNil(activePoint)) {
+          return [
+            graphicalItem,
+            ...this.renderActivePoints({
+              item, activePoint, basePoint, childIndex: activeTooltipIndex,
+              isRange,
+            }),
+          ];
+        }
       }
 
       if (isRange) { return [graphicalItem, null, null]; }