+++ src/component/Label.js
@@ -91,7 +91,7 @@ const renderRadialLabel = (labelProps, label, attrs) => {
   const path = `M${startPoint.x},${startPoint.y}
     A${radius},${radius},0,1,${direction ? 0 : 1},
     ${endPoint.x},${endPoint.y}`;
-  const id = uniqueId('recharts-radial-line-');
+  const id = _.isNil(labelProps.id) ? uniqueId('recharts-radial-line-') : labelProps.id;
 
   return (
     <text
@@ -290,7 +290,6 @@ function Label(props) {
     if (isValidElement(label)) {
       return label;
     }
-    console.log(label);
   } else {
     label = getLabel(props);
   }